# docs/swagger.yaml
definitions:
  models.Item:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
paths:
  /items/:
    get:
      description: Retrieves a list of all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Item'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all items
      tags:
      - Items
    post:
      consumes:
      - application/json
      description: Adds a new item to the database
      parameters:
      - description: Item Data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new item
      tags:
      - Items
  /items/{id}:
    delete:
      description: Deletes an item by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an item
      tags:
      - Items
    get:
      description: Retrieves an item by its ID
      parameters:
      - description: Item ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an item by ID
      tags:
      - Items
    put:
      consumes:
      - application/json
      description: Updates an existing item
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Item Data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an item
      tags:
      - Items
swagger: "2.0"